+++ CHANGELOG.md
@@ -3,6 +3,17 @@ CHANGELOG
 
 0.41.0
 ------
+- Added color name `preview-border` and `preview-scrollbar`
+- Added new border style `block` which uses [block elements](https://en.wikipedia.org/wiki/Block_Elements)
+- `--scrollbar` can take two characters, one for the main window, the other
+  for the preview window
+- Putting it altogether
+  ```sh
+  fzf-tmux -p 80% --padding 1,2 --preview 'bat --style=plain --color=always {}' \
+      --color 'bg:#222233,bg+:#333344,gutter:#222233,border:#111122,scrollbar:#ffaa00' \
+      --color 'preview-bg:#332233,preview-border:#222222,preview-scrollbar:#00aaff' \
+      --preview-window 'border-block' --border block --scrollbar '▌▐'
+  ```
 - Bug fixes and improvements
 
 0.40.0